Save work prior to cacophonous AgentProtocol introduction.
